+++ game/include/main/obj_message.h
@@ -5,8 +5,8 @@
 
 extern char sObjMsgOverflowInObjectWarning[];
 
-int ObjMsg_Peek(GameObject* obj, u32* outMessage, u32* outSender, u32* outParam);
-int ObjMsg_Pop(GameObject* obj, u32* outMessage, u32* outSender, u32* outParam);
+int ObjMsg_Peek(GameObject* obj, u32* outMessage, uintptr_t* outSender, u32* outParam);
+int ObjMsg_Pop(GameObject* obj, u32* outMessage, uintptr_t* outSender, u32* outParam);
 void ObjMsg_SendToNearbyObjects(int targetId, f32 radius, u32 flags, void* sender, u32 message, u32 param);
 void ObjMsg_SendToObjects(int targetId, u32 flags, void* sender, u32 message, u32 param);
 u32 ObjMsg_SendToObject(GameObject* obj, u32 message, void* sender, u32 param);
