+++ game/src/dlls/engine/0/0.c
@@ -4238,7 +4238,7 @@ void headDisplayDraw(void)
                       (f32)(u32)gRenderModeObj->xfbHeight, 0.0f, 1.0f);
         if (gHeadDisplayModelObjs[panelType] != NULL)
         {
-            ObjAnim_AdvanceCurrentMove((int)gHeadDisplayModelObjs[panelType], gPauseMenuPanelAnims.speeds[panelType], timeDelta, NULL);
+            ObjAnim_AdvanceCurrentMove(gHeadDisplayModelObjs[panelType], gPauseMenuPanelAnims.speeds[panelType], timeDelta, NULL);
             if (gHeadDisplayModelObjs[panelType]->anim.placementDataAddress > 0x90000000u)
             {
                 gHeadDisplayModelObjs[panelType]->anim.placementDataAddress = 0;
@@ -4389,7 +4389,7 @@ void pauseMenuCreateHeads(void)
                 {
                     ((GameObject*)gHeadDisplayModelObjs[i])->anim.placementData = NULL;
                 }
-                ObjAnim_SetCurrentMove((int)gHeadDisplayModelObjs[i], 1, 0.0f, 0);
+                ObjAnim_SetCurrentMove(gHeadDisplayModelObjs[i], 1, 0.0f, 0);
             }
         }
     }
@@ -6063,7 +6063,7 @@ void pauseMenuUpdate(void)
             pauseMenuSetupTitle(0x2b1, gPauseMenuPageIndex, 1, 3);
             if ((s8)gPauseMenuCloseAnimIndex != 0 && AudioStream_GetCurrentId() == 0 && AudioStream_IsPreparing() == 0)
             {
-                ObjAnim_SetCurrentMove((int)hud->anims[(s8)gPauseMenuCloseAnimIndex], 0, 0.0f, 0);
+                ObjAnim_SetCurrentMove(hud->anims[(s8)gPauseMenuCloseAnimIndex], 0, 0.0f, 0);
                 gPauseMenuCloseAnimIndex = 0;
             }
             if ((s8)analogX != 0 || gPauseMenuPodiumRamp == 0 || gPauseMenuPageIndex < menuMin || gPauseMenuPageIndex > menuMax)
@@ -6722,7 +6722,7 @@ void pauseMenuUpdateFadeAndBack(void)
         }
         for (i = 1; i < 4; i++)
         {
-            ObjAnim_SetCurrentMove((int)gGameUiHudAnimObjects[i], i == (s8)gPauseMenuCloseAnimIndex, 0.0f, 0);
+            ObjAnim_SetCurrentMove(gGameUiHudAnimObjects[i], i == (s8)gPauseMenuCloseAnimIndex, 0.0f, 0);
         }
     }
 
@@ -7030,7 +7030,7 @@ void pauseMenuAnimateCarousel(void)
         f32 spin = 0.13f * gPauseMenuPodiumRamp;
         gGameUiCommunicatorObjects[1]->anim.rootMotionScale = spin * gPauseMenuRingScale;
     }
-    ObjAnim_AdvanceCurrentMove((int)gGameUiCommunicatorObjects[1], 0.01f, timeDelta,
+    ObjAnim_AdvanceCurrentMove(gGameUiCommunicatorObjects[1], 0.01f, timeDelta,
                                &animEvents);
     watermark = 0x90000000;
     for (; k <= last; k++)
@@ -7053,7 +7053,7 @@ void pauseMenuAnimateCarousel(void)
         sel = gPauseMenuRingExpand * sel;
         gGameUiHudAnimObjects[k]->anim.rootMotionScale = sel * gPauseMenuRingScale;
         gGameUiHudAnimObjects[k]->anim.renderAlpha = 0xff;
-        ObjAnim_AdvanceCurrentMove((int)gGameUiHudAnimObjects[k], gPauseMenuPanelAnims.speeds[k], timeDelta,
+        ObjAnim_AdvanceCurrentMove(gGameUiHudAnimObjects[k], gPauseMenuPanelAnims.speeds[k], timeDelta,
                                                                      &animEvents);
         a = 2.0f * mathSinf(gGameUiPi * (f32)(gPauseMenuSwivelAngle + k * step) / gGameUiAngleDivisor);
         a = gPauseMenuRingExpand * a;
