+++ game/src/dlls/objects/203/203.c
@@ -179,7 +179,7 @@ int dll_CB_moveHandler1(GameObject* obj, GroundBaddieState* state, f32 timeDelta
     (void)timeDelta;
 
     if (state->baddie.moveJustStartedA != 0) {
-        ObjAnim_SetCurrentMove((int)obj, 0, 0.0f, 0);
+        ObjAnim_SetCurrentMove(obj, 0, 0.0f, 0);
         state->baddie.moveDone = 0;
     }
     state->baddie.physicsActive = 1;
